net/ixgbe: shrink size of context index
The ixgbe driver only supports two contexts, so using uint32_t for the
context index is excessive. Reducing this to uint8_t shrinks the ixgbe
part of the union from 32 bytes to 24. Although it does not in itself
shrink the space for the whole struct, it is worth doing as ixgbe is the
second-largest block in the union, so may become relevant if other
rework shrinks the idpf-specific block.

Acked-by: Anatoly Burakov <anatoly.burakov@intel.com>

+++ drivers/net/intel/common/tx.h
@@ -200,7 +200,7 @@ struct ci_tx_queue {
 		struct { /* ixgbe specific values */
 			const struct ixgbe_txq_ops *ops;
 			struct ixgbe_advctx_info *ctx_cache;
-			uint32_t ctx_curr;
+			uint8_t ctx_curr;
 			uint8_t pthresh;   /**< Prefetch threshold register. */
 			uint8_t hthresh;   /**< Host threshold register. */
 			uint8_t wthresh;   /**< Write-back threshold reg. */
